Fixed more the edges computation. This is the version used for paper.
We now set as endpoint not all the endpoint of the edges but rather the ones that are big enough so that we dont set too many point on vegetation areas and we still manage to reconstruct the part of the wall of the LBH that was wobbly before.
However this has a slight problem now with the poles. The solution would be to also set as endpoints those edges which are "valid" and surrounded by invalid edges. Validty is expressed as being at a too grayin angle. Therefore edges that lie on poles will be valid (almost orthogonal to the view) surrounded by non valid edges (almost paralel to the view).